requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
Pillow>=10.0
PyJWT>=2.9
python-dotenv>=1.0
